new_config -> config
