--- mbed_build/_internal/config/__init__.py
@@ -1,4 +1,0 @@
-#
-# Copyright (C) 2020 Arm Mbed. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
