--- mbed_build/_internal/config/config_modifiers/__init__.py
@@ -1,6 +1,0 @@
-#
-# Copyright (C) 2020 Arm Mbed. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-"""Public interface of config modifiers package."""
-from mbed_build._internal.config.config_modifiers.build_modifiers import *
